refactor migration structure and update version handling
